funçao cgeckIfIsAdm dentro do utils implementada, criaçao e organizaçao de rotas, erro no jwt nao estar setando o usuario logado corrigido, implementando getUsuarioPorHabilidade e adicionaHabilidadeAoPerfil

+++ controllers/habilidadesController.js
@@ -1,83 +1,103 @@
-const { Router } = require("express")
 const { Habilidades } = require("../models")
 const { habilidadeSchema } = require("../schemas/habilidadeSchema")
-const router = Router()
+const checkIfIsAdmin = require("../utils")
 
 //GET ALL
-router.get("/", async (req, res) => {
+const getAll = async function (req, res) {
     try {
-        console.log(req.body)
         const habilidades = await Habilidades.findAll()
-        res.status(200).json(habilidades)
+        return res.status(200).json(habilidades)
     } catch (erro) {
-        res.status(500).json(erro)
+        return res.status(500).json(erro)
     }
-})
+}
 
 //GET UMA SÓ HABILIDADE
-router.get("/:id", async (req, res) => {
+const getHabilidadeUnica = async function (req, res) {
     try {
         const habilidade = await Habilidades.findByPk(req.params.id) // Ou usar o findOne, mas não sei como ligar ele ao front pra pegar o input da habilidade
-        res.status(200).json(habilidade)
+        return res.status(200).json(habilidade)
     } catch (erro) {
-        res.status(500).json({message: erro})
+        return res.status(500).json({ message: erro })
     }
-})
+}
 
 //CREATE
-router.post("/", async (req, res) => {
+//IMPLEMENTAR ESSE METODO E O ALTERA E DELETA PRA SER EXCLUSIVO DE ADM, NO FROM, SE FOR USUARIO COMUM, ELE SÓ VAI PODER UTILIZAR O ADICIONA HABILIDADE AO PERFIL.
+
+//ESSE METODO É PRA CADASTRAR A HABILIDADE NO SISTEMA PROS USUARIOS PODEREM SELECIONAR ELA NO CHECKBOX E ADICIONAR AO SEU PERFIL
+const criaHabilidade = async function (req, res) {
     try {
+        if (!(await checkIfIsAdmin(req, res))) {
+            res.status(401)
+            throw new Error("Usuario não administrador")
+        }
+
         const data = req.body
 
         const habilidadeValida = await habilidadeSchema.validate(data)
-        console.log(habilidadeValida)
 
-        if (habilidadeValida.error) throw new Error(habilidadeValida.error)
+        if (habilidadeValida.error) {
+            res.status(400)
+            throw new Error(habilidadeValida.error)
+        }
 
         const existeNoSistema = await Habilidades.findOne({
             where: { nome: data.nome },
         })
-        if (existeNoSistema)
+        if (existeNoSistema) {
+            res.status(400)
             throw new Error("habilidade já cadastrada no sistema")
+        }
 
         if (habilidadeValida && !existeNoSistema) {
             const habilidadeCriada = await Habilidades.create(data)
 
-            res.status(200).json({
+            return res.status(200).json({
                 message: "Habilidade cadastrada com sucesso:",
                 habilidadeCriada: habilidadeCriada,
             })
         }
     } catch (erro) {
-        res.status(500).json({
+        return res.json({
             message: `Ocorreu um erro.. ${erro}`,
         })
-        console.log(erro)
     }
-})
+}
 
 //DELETE
-router.delete("/:id", async (req, res) => {
+const deletaHabilidade = async function (req, res) {
     try {
+        if (!(await checkIfIsAdmin(req, res))) {
+            res.status(401)
+            throw new Error("Usuario não administrador")
+        }
+
         const deletedHabilidade = await Habilidades.destroy({
             where: {
                 id: req.params.id,
             },
         })
-        res.status(200).json({
+        return res.status(200).json({
             message: "Habilidades deletada com sucesso",
             deletedHabilidade: req.body,
         })
     } catch (erro) {
-        res.status(500).json({
+        return res.status(500).json({
             message: erro,
         })
     }
-})
+}
 
 //UPDATE
-router.put("/:id", async (req, res) => {
+const alteraHabilidade = async function (req, res) {
     try {
+
+        if (!(await checkIfIsAdmin(req, res))) {
+            res.status(401)
+            throw new Error("Usuario não administrador")
+        }
+
         const { nome } = req.body
 
         const updatedHabilidade = await Habilidades.update(
@@ -90,13 +110,19 @@ router.put("/:id", async (req, res) => {
                 },
             }
         )
-        res.status(200).json({
+        return res.status(200).json({
             message: "Habilidade alterada com sucesso",
             updatedHabilidade: req.body,
         })
     } catch (erro) {
-        res.status(500).json({ message: erro })
+        return res.status(500).json({ message: erro })
     }
-})
+}
 
-module.exports = router
+module.exports = {
+    getAll,
+    getHabilidadeUnica,
+    criaHabilidade,
+    alteraHabilidade,
+    deletaHabilidade,
+}
